setup the buildVideoElement for the Video SDK
