+++ packages/js/src/Client.ts
@@ -8,10 +8,7 @@ import {
 } from '@signalwire/core'
 import type { CustomSaga } from '@signalwire/core'
 import { ConnectionOptions } from '@signalwire/webrtc'
-import {
-  makeVideoElementSaga,
-  makeAudioElementSaga,
-} from './features/mediaElements/mediaElementsSagas'
+import { makeAudioElementSaga } from './features/mediaElements/mediaElementsSagas'
 import {
   createBaseRoomSessionObject,
   RoomSessionConnection,
@@ -20,8 +17,9 @@ import { VideoManager, createVideoManagerObject } from './cantina'
 import type { Client as ChatClient } from './chat/Client'
 import type { Client as PubSubClient } from './pubSub/Client'
 import type { RoomSession } from './RoomSession'
+import { buildVideoElement } from './fabric'
 
-export interface Client<RoomSessionType = RoomSession>
+export interface Client<RoomSessionType extends RoomSession = RoomSession>
   extends ClientContract<Client<RoomSessionType>, ClientEvents> {
   rooms: ClientAPI<RoomSessionType>['rooms']
   chat: ClientAPI<RoomSessionType>['chat']
@@ -46,7 +44,7 @@ export interface MakeRoomOptions extends ConnectionOptions {
 }
 
 export class ClientAPI<
-  RoomSessionType = RoomSession
+  RoomSessionType extends RoomSession = RoomSession
 > extends BaseClient<ClientEvents> {
   private _videoManager: VideoManager
   private _chat: ChatClient
@@ -58,6 +56,8 @@ export class ClientAPI<
         const {
           rootElement,
           applyLocalVideoOverlay = true,
+          applyMemberOverlay = true,
+          mirrorLocalVideoOverlay = true,
           stopCameraWhileMuted = true,
           stopMicrophoneWhileMuted = true,
           ...options
@@ -76,24 +76,29 @@ export class ClientAPI<
           })
         )
 
-        /**
-         * If the user provides a `rootElement` we'll
-         * automatically handle the Video element for them
-         */
-        if (rootElement) {
-          customSagas.push(
-            makeVideoElementSaga({
-              rootElement,
-              applyLocalVideoOverlay,
-            })
-          )
-        }
-
         const room = createBaseRoomSessionObject<RoomSessionType>({
           ...options,
           store: this.store,
           customSagas,
         })
+
+        /**
+         * If the user provides a `rootElement` we'll
+         * automatically handle the Video element for them
+         */
+        if (rootElement) {
+          try {
+            buildVideoElement({
+              applyLocalVideoOverlay,
+              applyMemberOverlay,
+              mirrorLocalVideoOverlay,
+              room,
+              rootElement,
+            })
+          } catch (error) {
+            this.logger.error('Unable to build the video element automatically')
+          }
+        }
 
         /**
          * If the user joins with `join_video_muted: true` or
